% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/accuracyTableVis.R
\name{accuracyTableVis}
\alias{accuracyTableVis}
\title{Create an accuracy table visualization}
\usage{
accuracyTableVis(plot.title = "Sensitivity/Specificity Plot",
  colors = c("deepskyblue", "orangered"))
}
\arguments{
\item{plot.title}{The title of the plot to be created}

\item{colors}{The colors used (two colors, first for TP/TN)}
}
\description{
\code{accuracyTableVis} allows users to examine the effect of changing
a test's sensitivity and specificity or adjusting the prevalence of the
disease being tested. Changes in the number of true positives (TP), false
positives (FP), true negatives (TN), and false negatives (FN)
are reactively displayed.
}
\examples{
accuracyTableVis()
}

