% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buildWidget.R
\name{buildWidget}
\alias{buildWidget}
\title{Create an HTML widget}
\usage{
buildWidget(widget_name, data, settings, width = NULL, height = NULL)
}
\arguments{
\item{widget_name}{Character string indicating what type of widget to build}

\item{data}{Data as specified by Vis function}

\item{settings}{Settings as specified by Vis function}
}
\description{
Create an HTML widget
}

