% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/iconArrayVis.R
\name{iconArrayVis}
\alias{iconArrayVis}
\title{Create an icon array.}
\usage{
iconArrayVis(mobj = NULL, data = NULL, group.colors = c("deepskyblue",
  "orangered"), init.color = "lightgray", plot.title = "Icon Array")
}
\arguments{
\item{mobj}{Result of multinomial logit regression. Defaults to NULL, meaning groups are manually specified.}

\item{colors}{List of colors to use for each group. Should be length of levels of outcome or length of group.names.}
}
\description{
\code{iconArrayVis} creates an interactive icon array. Allows for specification of a multinomial
logit regression to estimate probability of each group based on a set of covariates. This concept is
based on research by Zikmund-Fisher et. al. (http://www.iconarray.com/)
}
\examples{
# To create a simple icon array image
iconArrayVis()
# To display the results of a multinomial logit (quine data from MASS package)
library(MASS)
library (nnet) # contains function multinom
mobj <- multinom(Age~Eth+Sex+Lrn+Days, data=quine)
iconArrayVis(mobj, data=quine, colors=c("deepskyblue", "orangered"), plot.title="School Absenteeism")
}

