\name{initialize}
\alias{initialize}
\title{Constructor for healthvis objects}
\arguments{
  \item{plotType}{the type of plot}

  \item{plotTitle}{the title of the plot}

  \item{varType}{a list of variable types (list contains
  "continuous" or "factor")}

  \item{d3Params}{list of parameters used in d3 script}

  \item{gaeDevel}{Use the appengine local dev server (for
  testing only, users should ignore)}
}
\description{
  Posts data to server
}

