% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/listHealthVis.R
\name{listHealthVis}
\alias{listHealthVis}
\title{This function lists all the currently available healthvis plots}
\usage{
listHealthVis()
}
\description{
This function lists all the currently available healthvis plots
}

