\name{plot.healthvis}
\alias{plot.healthvis}
\title{A method for plotting a healthvis object}
\usage{
  \method{plot}{healthvis} (x,
    file.name = "healthvis.html")
}
\arguments{
  \item{x}{A healthvis object to be plotted}

  \item{file.title}{The file name for the HTML page to be
  created}
}
\value{
  \code{NULL}
}
\description{
  A method for plotting a healthvis object
}

