\name{writeD3AccuracyTable}
\alias{writeD3AccuracyTable}
\title{This function writes the D3.js code for the accuracy table graphic.}
\usage{
  writeD3AccuracyTable()
}
\value{
  d3.js The D3.js javascript implementing the plot
}
\description{
  This function writes the D3.js code for the accuracy
  table graphic.
}

