\name{writeD3AccuracyTableCss}
\alias{writeD3AccuracyTableCss}
\title{This function writes the CSS code for the accuracy table SVG.}
\usage{
  writeD3AccuracyTableCss(line.size = 2, axis.size = 1)
}
\arguments{
  \item{line.size}{The line width}

  \item{axis.size}{The axis width}
}
\value{
  d3.css The CSS for the D3.js implemented SVG.
}
\description{
  This function writes the CSS code for the accuracy table
  SVG.
}

