\name{writeD3SurvivalCss}
\alias{writeD3SurvivalCss}
\title{This function writes the CSS code for the survival SVG.}
\usage{
  writeD3SurvivalCss(line.size = 2, axis.size = 1)
}
\arguments{
  \item{line.size}{The line width}

  \item{axis.size}{The axis width}
}
\value{
  d3.css The CSS for the D3.js implemented SVG.
}
\description{
  This function writes the CSS code for the survival SVG.
}

