\name{accuracyTableVis}
\alias{accuracyTableVis}
\title{This function creates an interactive Accuracy Table visualization}
\usage{
  accuracyTableVis(plot.title = "Sensitivity/Specificity Plot",
    plot = TRUE)
}
\arguments{
  \item{plot.title}{The title of the plot to be created}

  \item{plot}{If TRUE the plot is launched in a browser.}
}
\description{
  \code{accuracyTableVis} creates an interactive survival
  plot
}

